requirements.txt
pelican==4.11.0
Markdown==3.7
PyYAML==6.0.1